Fixes #2269. Backports fix for deleting video sources in mythtv-setup.
When you deleted a video source, mythtv-setup left behind the channels and the capture card input connections to the source. This extends the deletion to remove this associated data as well.



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-20-fixes@11425 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/videosource.cpp
@@ -2079,6 +2079,30 @@ void VideoSourceEditor::del()
     if (val == 0)
     {
         MSqlQuery query(MSqlQuery::InitCon());
+
+        // Delete the channels associated with the source
+        query.prepare("DELETE FROM channel "
+                      "WHERE sourceid = :SOURCEID");
+        query.bindValue(":SOURCEID", getValue());
+
+        if (!query.exec() || !query.isActive())
+        {
+            MythContext::DBError("Deleting Channels", query);
+            return;
+        }
+
+        // Delete the inputs associated with the source
+        query.prepare("DELETE FROM cardinput "
+                      "WHERE sourceid = :SOURCEID");
+        query.bindValue(":SOURCEID", getValue());
+
+        if (!query.exec() || !query.isActive())
+        {
+            MythContext::DBError("Deleting cardinputs", query);
+            return;
+        }
+
+        // Delete the source itself
         query.prepare("DELETE FROM videosource "
                       "WHERE sourceid = :SOURCEID");
         query.bindValue(":SOURCEID", getValue());
